refactor: folders/files architecture
