+++ client/package.json
@@ -23,7 +23,7 @@
     "react-to-print": "^2.14.12"
   },
   "scripts": {
-    "start": "react-scripts start ",
+    "start": "BROWSER=none react-scripts start ",
     "build": "react-scripts build && echo '/* /index.html 200' | cat >build/_redirects",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
