+++ client/src/components/guards/GuestRoute.jsx
@@ -1,9 +1,9 @@
 import { useContext } from 'react';
 import { AuthContext } from '../../context/auth.context';
 import { Navigate } from 'react-router-dom';
-import { Loading } from '../../components';
+import { Loading } from '..';
 
-export function IsPrivate({ children }) {
+export function GuestRoute({ children }) {
   const { isLoggedIn, isLoading } = useContext(AuthContext);
 
   // If the authentication is still loading ⏳
@@ -11,10 +11,11 @@ export function IsPrivate({ children }) {
     return <Loading />;
   }
 
-  if (!isLoggedIn) {
-    // If the user is not logged in navigate to the login page ❌
-    return <Navigate to="/login" />;
+  if (isLoggedIn) {
+    // If the user is logged in, navigate to home page ❌
+    return <Navigate to="/" />;
   }
-  // If the user is logged in, allow to see the page ✅
+
+  // If the user is not logged in, allow to see the page ✅
   return children;
 }
