+++ client/src/components/guards/PrivateRoute.jsx
@@ -1,9 +1,9 @@
 import { useContext } from 'react';
 import { AuthContext } from '../../context/auth.context';
 import { Navigate } from 'react-router-dom';
-import { Loading } from '../../components';
+import { Loading } from '..';
 
-export function IsAnon({ children }) {
+export function PrivateRoute({ children }) {
   const { isLoggedIn, isLoading } = useContext(AuthContext);
 
   // If the authentication is still loading ⏳
@@ -11,11 +11,10 @@ export function IsAnon({ children }) {
     return <Loading />;
   }
 
-  if (isLoggedIn) {
-    // If the user is logged in, navigate to home page ❌
-    return <Navigate to="/" />;
+  if (!isLoggedIn) {
+    // If the user is not logged in navigate to the login page ❌
+    return <Navigate to="/login" />;
   }
-
-  // If the user is not logged in, allow to see the page ✅
+  // If the user is logged in, allow to see the page ✅
   return children;
 }
