+++ client/src/components/index.js
@@ -1,31 +1,33 @@
-export { CreateAboutMe } from './AboutMe/CreateAboutMe';
-export { DeleteAboutMe } from './AboutMe/DeleteAboutMe';
+export { CreateAboutMe } from './pages/aboutMePage/CreateAboutMe';
+export { DeleteAboutMe } from './pages/aboutMePage/DeleteAboutMe';
 
-export { CreateAPost } from './BlogPost/CreateAPost';
-export { modules } from './BlogPost/EditorToolbar';
-export { EditPostCard } from './BlogPost/EditPostCard';
-export { LikeButton } from './BlogPost/LikeButton';
-export { PostCard } from './BlogPost/PostCard';
-export { CommentBox } from './CommentBox/CommentBox';
+export { CreateATopic } from './pages/postsPage/topic/CreateATopic';
+export { CreateAPost } from './pages/postsPage/CreateAPost';
+export { EditTopic } from './pages/postsPage/topic/EditTopic';
+export { TopicsSection } from './pages/postsPage/topic/TopicsSection';
+export { PostResults } from './pages/postsPage/PostResults';
+export { PostCard } from './pages/postsPage/PostCard';
 
-export { AddPicture } from './Gallery/AddPicture';
-export { GalleryPage } from './Gallery/GalleryPage';
-export { EditPicture } from './Gallery/EditPicture';
-export { DeletePicture } from './Gallery/DeletePicture';
+export { modules } from './pages/singlePostPage/EditorToolbar';
+export { EditPostCard } from './pages/singlePostPage/EditPostCard';
+export { LikeButton } from './pages/singlePostPage/LikeButton';
+export { CommentBox } from './pages/singlePostPage/CommentBox';
 
-export { IsAnon } from './IsAnon/IsAnon';
-export { IsPrivate } from './IsPrivate/IsPrivate';
-export { Loading } from './Loading/Loading';
-export { Navbar } from './Navbar/Navbar';
+export { AddPicture } from './pages/galleryPage/AddPicture';
+export { GalleryPage } from './pages/galleryPage/GalleryPage';
+export { EditPicture } from './pages/galleryPage/EditPicture';
+export { DeletePicture } from './pages/galleryPage/DeletePicture';
 
-export { CreateATopic } from './Topic/CreateATopic';
-export { EditTopic } from './Topic/EditTopic';
-export { TopicsSection } from './Topic/TopicsSection';
+export { AddWords } from './pages/wordsPage/AddWords';
+export { DeleteWords } from './pages/wordsPage/DeleteWords';
+export { EditWords } from './pages/wordsPage/EditWords';
+export { ReadingTime } from './pages/wordsPage/ReadingTime';
+export { Words } from './pages/wordsPage/Words';
 
-export { PostResults } from './Posts/PostResults';
+export { GuestRoute } from './guards/GuestRoute';
+export { PrivateRoute } from './guards/PrivateRoute';
 
-export { AddWords } from './Words/AddWords';
-export { DeleteWords } from './Words/DeleteWords';
-export { EditWords } from './Words/EditWords';
-export { ReadingTime } from './Words/ReadingTime';
-export { Words } from './Words/Words';
+export { Loading } from './ui/loading/Loading';
+export { Navbar } from './ui/navbar/Navbar';
+
+export { MetaData } from './common/Metadata';
