+++ client/src/components/pages/galleryPage/AddPicture.jsx
