+++ client/src/components/pages/galleryPage/DeletePicture.jsx
