+++ client/src/components/pages/galleryPage/EditPicture.jsx
