+++ client/src/components/pages/wordsPage/DeleteWords.jsx
