+++ client/src/components/pages/wordsPage/ReadingTime.jsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export const ReadingTime = ({ text }) => {
   const wordsPerMinute = 200;
   const words = text.split(/\s+/g).length;
