+++ client/src/components/ui/loading/Loading.css
@@ -1,20 +1,17 @@
 .wrapper {
   height: 100vh;
   width: 100vw;
-  display: flex;
-  justify-content: center;
-  align-items: center;
   background: none;
 }
 
 .spinner {
   width: 300px;
-  text-align: center;
+  text-align: start;
 }
 
 .spinner > div {
-  width: 36px;
-  height: 36px;
+  width: 18px;
+  height: 18px;
   background-color: #333;
 
   border-radius: 100%;
