+++ client/src/pages/ContactMePage.jsx
@@ -1,7 +1,7 @@
 import React, { useRef } from 'react';
 import emailjs from '@emailjs/browser';
 import { useNavigate } from 'react-router-dom';
-import MetaData from '../components/Metadata';
+import { MetaData } from '../components';
 
 export const ContactMePage = () => {
   const navigate = useNavigate();
